choice in formdetailview

+++ form_builder/forms/views.py
@@ -8,6 +8,7 @@ from django.contrib.auth.hashers import check_password
 from django.db.models import Q
 from django.shortcuts import reverse
 from rest_framework import viewsets, status
+from rest_framework.views import APIView
 from rest_framework.decorators import action
 from rest_framework.response import Response
 from django.http import Http404
@@ -16,7 +17,7 @@ from .serializers import UserRetrievalSerializer,FormListSerializer,FormCUDSeria
 from user.models import User
 from .models import Form,FormField
 from utils.permission import JWTUtils
-
+from utils.response import CustomResponse
 
 # Create your views here.
 
@@ -33,7 +34,7 @@ class FormViewSet(viewsets.ModelViewSet):
     
     def get_queryset(self):
         user_id = JWTUtils.fetch_user_id(self.request)
-        return Form.objects.filter(user=user_id)
+        return Form.objects.filter(user_id=user_id)
     
     def perform_create(self, serializer):
         user_id = JWTUtils.fetch_user_id(self.request)
@@ -42,6 +43,52 @@ class FormViewSet(viewsets.ModelViewSet):
     def perform_update(self, serializer):
         user_id = JWTUtils.fetch_user_id(self.request)
         serializer.save(user_id=user_id)
+
+
+    def list(self, request, *args, **kwargs):
+        try:
+            queryset = self.filter_queryset(self.get_queryset())
+            serializer = self.get_serializer(queryset, many=True)
+            
+            return CustomResponse(response=serializer.data).get_success_response()
+        except Exception as e:
+            return CustomResponse(message=str(e)).get_failure_response()
+
+    def retrieve(self, request, *args, **kwargs):
+        try:
+            instance = self.get_object()
+            serializer = self.get_serializer(instance)
+            return CustomResponse(response=serializer.data).get_success_response()
+        except Exception as e:
+            return CustomResponse(message=str(e)).get_failure_response()
+
+    def create(self, request, *args, **kwargs):
+        serializer = self.get_serializer(data=request.data)
+        if serializer.is_valid():
+            self.perform_create(serializer)
+            return CustomResponse(response=serializer.data).get_success_response()
+        else:
+            return CustomResponse(message=serializer.errors).get_failure_response()
+    
+    def update(self, request, *args, **kwargs):
+        instance = self.get_object()
+        serializer = self.get_serializer(instance, data=request.data, partial=True)
+        if serializer.is_valid():
+            self.perform_update(serializer)
+            return CustomResponse(response=serializer.data).get_success_response()
+        else:
+            return CustomResponse(message=serializer.errors).get_failure_response()
+    
+    def destroy(self, request, *args, **kwargs):
+        try:
+            instance = self.get_object()
+            self.perform_destroy(instance)
+            return CustomResponse(message="Form deleted successfully").get_success_response()
+        except Http404:
+            return CustomResponse(message="Form not found").get_failure_response(status_code=status.HTTP_404_NOT_FOUND)
+        except Exception as e:
+            return CustomResponse(message=str(e)).get_failure_response(status_code=status.HTTP_400_BAD_REQUEST)
+
     
     @action(detail=True, methods=['post'])
     def add_field(self, request, pk=None):
@@ -49,32 +96,33 @@ class FormViewSet(viewsets.ModelViewSet):
         try:
             form = Form.objects.get(pk=pk, user_id=user_id)
         except Form.DoesNotExist:
-            raise Http404("Form does not exist")
+            return CustomResponse(message="form not found").get_failure_response(status_code=status.HTTP_404_NOT_FOUND)
         
         serializer = FormFieldSerializer(data=request.data, context={'form': form})
         if serializer.is_valid():
             serializer.save()
-            return Response({"message": "Form field added"}, status=status.HTTP_201_CREATED)
+            return CustomResponse(message="Formfield added").get_success_response()
         else:
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            return CustomResponse(message=serializer.errors).get_failure_response()
     
     @action(detail=True, methods=['put'], url_path='edit_field/(?P<field_pk>[^/.]+)')
     def edit_field(self, request, pk=None, field_pk=None):
         user_id = JWTUtils.fetch_user_id(request)
+        
         try:
             formfield = FormField.objects.select_related('form').get(pk=field_pk)
         except FormField.DoesNotExist:
-            raise Http404("Form field does not exist")
+            return CustomResponse(message="formfield does not exits").get_failure_response(status_code=status.HTTP_404_NOT_FOUND)
         
-        if formfield.form.user_id != user_id:
-            return Response({"error": "You do not have permission to edit this form field"}, status=status.HTTP_403_FORBIDDEN)
+        if formfield.form.user_id != user_id or formfield.form.id != pk:
+            return CustomResponse(message="formfield does not exits").get_failure_response(status_code=status.HTTP_404_NOT_FOUND)
         
         serializer = FormFieldSerializer(formfield, data=request.data)
         if serializer.is_valid():
             serializer.save()
-            return Response({"message": "Form field edited"}, status=status.HTTP_201_CREATED)
+            return CustomResponse(message="formfield edited").get_success_response()
         else:
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            return CustomResponse(message=serializer.errors).get_failure_response()
     
     @action(detail=True, methods=['delete'], url_path='delete_field/(?P<field_pk>[^/.]+)')
     def delete_field(self, request, pk=None, field_pk=None):
@@ -83,10 +131,31 @@ class FormViewSet(viewsets.ModelViewSet):
         try:
             formfield = FormField.objects.select_related('form').get(pk=field_pk)
         except FormField.DoesNotExist:
-            raise Http404("Form field does not exist")
+            return CustomResponse(message="Form field does not exist").get_failure_response()
         
-        if formfield.form.user_id != user_id:
-            return Response({"error": "You do not have permission to delete this form field"}, status=status.HTTP_403_FORBIDDEN)
+        if formfield.form.user_id != user_id or formfield.form.id != pk:
+            return CustomResponse(message="Form field does not exist").get_failure_response()
         
         formfield.delete()
-        return Response({"message": "Form field deleted successfully"}, status=status.HTTP_204_NO_CONTENT)
+        return CustomResponse(message="formfield deleted successfully").get_success_response()
+    
+
+
+class FormResponseAPI(APIView):
+    
+    def get(self, request, pk=None):
+        if not pk:
+            return CustomResponse(message="missing formID").get_failure_response()
+        try:
+            form = Form.objects.get(pk=pk)
+        except Form.DoesNotExist:
+            return CustomResponse(message="Form does not exits").get_failure_response(status_code=status.HTTP_404_NOT_FOUND)
+
+        serializer = FormDetailSerializer(form)
+        return Response(serializer.data)
+    
+    def post(self,request,pk = None):
+        if not pk:
+            return CustomResponse(message="missing formID").get_failure_response()
+        
+
